Now Refactoring - 1.
