+++ simmilarity.py
@@ -13,8 +13,11 @@ def init_directory(same_image_dir, nonsame_image_dir):
 def get_filenames(working_dir):
     # Get files and add dirname front of files.
     files = os.listdir(working_dir)
-    files = [working_dir + "/" + files[i] for i in range(len(files))]
-    return files
+    return_files = []
+    for i in range(len(files)):
+        if "." in files[i]:
+            return_files.append(working_dir + "/" + files[i])
+    return return_files
 
 def make_dir(dir_name):
     if not os.path.isdir(dir_name):
@@ -25,9 +28,9 @@ def get_not_same_file_name(nonsame_dir, i, j, file_path):
 
 
 if __name__ == "__main__":
-    working_dir = "./Image/Images"
-    same_dir = "./Image/Same"
-    nonsame_dir = "./Image/NotSame"
+    working_dir = "./Image/InteriorImage/livingroom - Big"
+    same_dir = working_dir + "/Same"
+    nonsame_dir =working_dir + "/NotSame"
     
     init_directory(same_dir, nonsame_dir)
     files = get_filenames(working_dir)
@@ -36,7 +39,10 @@ if __name__ == "__main__":
     simmilarity_list = []
     simmilarity_TF = [True for _ in range(file_amount)]
 
+    print("Searching TF_MAP ... ")
     for i in range(0, file_amount):
+        if i % (file_amount // 100) == 0:
+            print("=", end="")
         for j in range(i + 1, file_amount):
             score = subprocess.check_output(['pyssim', files[i], files[j]], encoding="utf8")
             if float(score) > 0.9:
@@ -46,7 +52,9 @@ if __name__ == "__main__":
             if DEBUG:
                 print("For " + files[i] + " and " + files[j])
                 print(score.rstrip())
-    
+    print()
+    print("Start for copying")
+
     # Copy to nonsame directory
     for i in range(0, file_amount):
         if simmilarity_TF[i]:
@@ -59,3 +67,4 @@ if __name__ == "__main__":
                 if simmilarity_list[j][0] == i:
                     shutil.copy2(files[simmilarity_list[j][0]], get_not_same_file_name(same_dir, i, j, files[simmilarity_list[j][0]]))
                     shutil.copy2(files[simmilarity_list[j][1]], get_not_same_file_name(same_dir, i, j, files[simmilarity_list[j][1]]))
+
